// web/src/generated/api/models/chatConfirmPart.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */
import type { ChatAction } from './chatAction';
import type { ChatConfirmPartTone } from './chatConfirmPartTone';
import type { ChatConfirmPartType } from './chatConfirmPartType';

export interface ChatConfirmPart {
  type: ChatConfirmPartType;
  label: string;
  description: string;
  tone: ChatConfirmPartTone;
  action: ChatAction;
}